+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.1
+
+- Bugfix(core/adapters): Fix TS issue with createEntityAdapter alt id
+
 ## 1.2.0
 
 - Refactor(core): Clean up types and allow sources factory